+++ packages/jsx-tokenizer/src/index.ts
@@ -1,12 +1,13 @@
-import { Accessor, Component, createComponent, createMemo, JSX } from "solid-js";
-import { isDev } from "solid-js/web";
+import { Accessor, Component, createComponent, createMemo, JSX, DEV } from "solid-js";
+import { isServer } from "solid-js/web";
 import type { ResolvedJSXElement, Narrow, Many } from "@solid-primitives/utils";
 import { asArray } from "@solid-primitives/utils";
 
 /** @internal $TYPE is only used for type inference */
 declare const $TYPE: unique symbol;
 /** @internal */
-const $TOKENIZER = Symbol(isDev ? "jsx-tokenizer" : "");
+// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+const $TOKENIZER = Symbol(!isServer && DEV ? "jsx-tokenizer" : "");
 
 /**
  * Identifies a JSX Tokenizer. It is returned by {@link createTokenizer} (or {@link createToken}) and used by {@link createToken} and {@link resolveTokens}.
@@ -58,7 +59,8 @@ export type TokenComponent<TProps extends object, TData = TProps> = Component<TP
  */
 export function createTokenizer<T>(options?: { name: string }): JSXTokenizer<T> {
   return {
-    [$TOKENIZER]: Symbol(isDev ? options?.name || "jsx-tokenizer" : ""),
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    [$TOKENIZER]: Symbol(!isServer && DEV ? options?.name || "jsx-tokenizer" : ""),
   } as JSXTokenizer<T>;
 }
 
@@ -100,14 +102,17 @@ export function createToken<P extends object, T>(
 ): TokenComponent<P, T> {
   const symbol =
     (args[0]?.[$TOKENIZER] ? (args.shift()[$TOKENIZER] as symbol) : undefined) ??
-    Symbol(isDev ? args[0]?.name || args[1]?.name || "jsx-token" : "");
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    Symbol(!isServer && DEV ? args[0]?.name || args[1]?.name || "jsx-token" : "");
 
   const comp = ((props: P) => {
     const token = (
       args[1]
         ? () => createComponent(args[1], props)
         : () => {
-            isDev &&
+            !isServer &&
+              // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+              DEV &&
               // eslint-disable-next-line no-console
               console.warn(
                 `Tokens can only be rendered with resolveTokens. ("${symbol.description}")`,
@@ -142,7 +147,8 @@ function getResolvedTokens(
       getResolvedTokens(resolved, value[i], symbols, addElements);
   // other element
   else if (addElements) resolved.push(value);
-  else if (isDev && value)
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+  else if (!isServer && DEV && value)
     // eslint-disable-next-line no-console
     console.warn(`Invalid JSX Element passed to token resolver:`, value);