+++ packages/range/src/indexRange.ts
@@ -1,5 +1,5 @@
-import { Accessor, createRoot, createSignal, onCleanup, Setter, untrack } from "solid-js";
-import { isDev } from "solid-js/web";
+import { Accessor, createRoot, createSignal, onCleanup, Setter, untrack, DEV } from "solid-js";
+import { isServer } from "solid-js/web";
 import { abs, ceil, min, RangeProps, sign, toFunction, accessor } from "./common";
 
 /**
@@ -103,8 +103,8 @@ export function indexRange<T>(
   return () => {
     let _step = getStep();
     if (_step === 0) {
-      // eslint-disable-next-line no-console
-      if (isDev) console.warn("Range cannot have a step of 0");
+      // eslint-disable-next-line no-console, @typescript-eslint/no-unnecessary-condition
+      if (!isServer && DEV) console.warn("Range cannot have a step of 0");
       return items;
     }
     const _start = getStart();